Simplify selector creation
This makes it possible to create a matrix that can be used to select rows and
columns from another matrix in a single call. Previously, this matrix needed to
be populated with a for-loop. Here's the new syntax:

```
auto selector = Matrix<size_t>(range(num), ys);
```

The first argument is a `std::vector<size_t>` that specifices which rows to
select. In this case, it's 0..num.

The second argument is a `std::vector<size_t` that specifices which columns to
select.

`num` is the number of row