Replace {lock,make}_weak with generic function
The previous versions of lock_weak and make_weak only worked on lists of
two and three arguments. Furthermore, the functions were long and
difficult to read.

By using variadic templates and arguments, these functions are now
generic in the number of arguments that they can take and are much
easier to read.