Attempt to fix divide() backward function
The divide() backward function is not passing tests. It's not clear if
the test is wrong or if the implementation is wrong.

This implementation uses the quotient rule to implement the backward
pass. Tests are still not passing.

Because this implementation is easier to read and appears to be correct,
I'm checking it in.

But because none of the examples use this functionality, I will revisit
later.