Fix bug in select_rows_and_cols backward pass
The `idx` value is captured weakly by the lambda, but is not captured in
the `prev` list. Because of this, the reference count for `idx` can go
to zero and be freed before use.

Luckily, this was caught by checking the weak pointer before use, so no
unsafe memory accesses were made.

However, this fix introduces a memory leak but capturing `idx` by value
in the lambda.