Add mlp example
This adds an example that is based on [makemore, part
2](https://www.youtube.com/watch?v=TCH_1BHY58I).

The dataset is split into three parts, train, dev, and test. During
training, an eval step that evaluates performance on the dev set is
performed.

At the end of training, the test set loss is evaluated, and then 20 new
names are generated.