Generalize and time mlp example
The mlp example now has three hyperparameters available when
constructing the model:

* context_size
* embedding_dim
* hidden

Also, a bug has been fixed in the mlp example. During the eval and test
phase, the backward pass was still called, causing the model to train on
the dev and test data. A gating flag has been added to the backward pass
to keep this from happening.

Timing for each training step have been added for benchmarking.