Add OpenMP support and safer memory accesses
In an effort to speed up training, OpenMP support has been added. This
results in a 3x speedup on the test system.

Also, in an effort to avoid out-of-bounds memory access, data access is
controlled in many places in storage.hpp by views.

These views are guaranteed to have have begin() and end() within the
underlying matrix storage. By using these iterators in ranged for loops,
along with `zip()`, access to the underlying data are guaranteed to be
in range.

This technique has not been applied in all places yet, but the longer
term goal is use this method everywhere.