Add MacOS build and run instructions, remove test_iterator
Building and running on MacOS requires g++, since clang does not support c++23
fully. Also, the Accelerate framework's gemm appeared to be quite slow in
comparison to OpenBLAS.

Because MacOS is not a priority, this was not investigated any further, and
instead instructions were added for installing OpenBLAS and g++ via Homebrew.

Also, a temporary test, called test_iterator, was added to test/CMakeLists.txt.
This should not have been added and was causing cmake to fail to generate the
build environment. This test has been removed.

Additionally, it was found that `climits.h` needed to be included in
`one_layer.cpp`. This has been added.