Add optimized BatchNorm1D
This implements an optimized version of BatchNorm1D. The number of
floating point operations for forward should be exactly the same, but
the number of memory accesses are dramatically lower.

The backward pass is both mathematically optimized and memory access
optimized.

The results are not exactly the same as PyTorch's BatchNorm1D, but
are fairly close and mlp_bn produces believable results.

Training for mlp_bn is faster by more than 2x.